Big Field Refactor in-progress
Batch of changes so far-- pushing so samples reflect changes.

Re-factoring the language model & compiler into associative maps
with fields, created all at once.

FOR THE USER:
This reinforces the idea that all source is just one constant value
(with braces and fields), makes syntax much easier to read, and
encourages decomposing templates using modules.

FOR THE COMPILER:
This guarantees that each module is a bag of symbol definitions,
making queries on modules easier.

WIP:
- See `src/pdm/parser/parser.yy`, `sandbox/`, and `eg/` for most recent
  reflection of language's design.
- Rest of code builds, but is far from correct or complete.
  - TODO: Complete 'stubs' that assert false until
    `pass3_emit_all NotImplemented` stub-- means typer OK.
  - TODO: Re-visit 'scoper' to handle module addresses
  - TODO: Post-typer, implement a variety of checks:
        1. expand templates into monomorphic variants
        2. check constants, esp. 'const' statement
        3. ensure value initialization not order-dependent
            - can be done while trying to evaluate constants to bake
              into field values.
- Probably more, I'm rusty :(